skin_ai_project/requirements_gradio.txt
gradio==4.44.1
gradio_client==1.3.0
huggingface_hub<1.0
pydantic<2.10
--extra-index-url https://download.pytorch.org/whl/cpu
torch
timm
albumentations
numpy
pillow